fix(search): fix default entities for aggregation filters
* Co-located aggregation functions in AggregationQueryBuilder
* Default list of entity specs for aggregation should include query by default field (at least 1)
* Improved error message for field Display Name collisions
