+++ metadata-io/src/main/java/com/linkedin/metadata/search/elasticsearch/query/request/AggregationQueryBuilder.java
@@ -1,36 +1,71 @@
 package com.linkedin.metadata.search.elasticsearch.query.request;
 
 import static com.linkedin.metadata.Constants.*;
+import static com.linkedin.metadata.search.utils.ESUtils.toFacetField;
 import static com.linkedin.metadata.utils.SearchUtil.*;
 
+import com.linkedin.data.template.LongMap;
 import com.linkedin.metadata.config.search.SearchConfiguration;
+import com.linkedin.metadata.models.EntitySpec;
 import com.linkedin.metadata.models.StructuredPropertyUtils;
 import com.linkedin.metadata.models.annotation.SearchableAnnotation;
+import com.linkedin.metadata.query.filter.ConjunctiveCriterion;
+import com.linkedin.metadata.query.filter.ConjunctiveCriterionArray;
+import com.linkedin.metadata.query.filter.Criterion;
+import com.linkedin.metadata.query.filter.CriterionArray;
+import com.linkedin.metadata.query.filter.Filter;
+import com.linkedin.metadata.search.AggregationMetadata;
+import com.linkedin.metadata.search.FilterValueArray;
 import com.linkedin.metadata.search.utils.ESUtils;
+import com.linkedin.metadata.utils.SearchUtil;
+import com.linkedin.util.Pair;
+import io.opentelemetry.extension.annotations.WithSpan;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
+import java.util.function.BinaryOperator;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang.StringUtils;
+import org.opensearch.action.search.SearchResponse;
+import org.opensearch.search.aggregations.Aggregation;
 import org.opensearch.search.aggregations.AggregationBuilder;
 import org.opensearch.search.aggregations.AggregationBuilders;
+import org.opensearch.search.aggregations.Aggregations;
+import org.opensearch.search.aggregations.bucket.missing.ParsedMissing;
+import org.opensearch.search.aggregations.bucket.terms.ParsedTerms;
+import org.opensearch.search.aggregations.bucket.terms.Terms;
 
 @Slf4j
 public class AggregationQueryBuilder {
+  private static final String URN_FILTER = "urn";
 
-  private final SearchConfiguration _configs;
-  private final Set<String> _defaultFacetFields;
-  private final Set<String> _allFacetFields;
+  private final SearchConfiguration configs;
+  private final Set<String> defaultFacetFields;
+  private final Set<String> allFacetFields;
+  private final Map<EntitySpec, List<SearchableAnnotation>> entitySearchAnnotations;
+
+  private Map<String, String> filtersToDisplayName;
 
   public AggregationQueryBuilder(
       @Nonnull final SearchConfiguration configs,
-      @Nonnull final List<SearchableAnnotation> annotations) {
-    this._configs = Objects.requireNonNull(configs, "configs must not be null");
-    this._defaultFacetFields = getDefaultFacetFields(annotations);
-    this._allFacetFields = getAllFacetFields(annotations);
+      @Nonnull Map<EntitySpec, List<SearchableAnnotation>> entitySearchAnnotations) {
+    this.configs = Objects.requireNonNull(configs, "configs must not be null");
+    this.entitySearchAnnotations = entitySearchAnnotations;
+
+    List<SearchableAnnotation> annotations =
+        this.entitySearchAnnotations.values().stream()
+            .flatMap(List::stream)
+            .collect(Collectors.toList());
+    this.defaultFacetFields = getDefaultFacetFields(annotations);
+    this.allFacetFields = getAllFacetFields(annotations);
   }
 
   /** Get the set of default aggregations, across all facets. */
@@ -48,7 +83,7 @@ public class AggregationQueryBuilder {
       facetsToAggregate =
           facets.stream().filter(this::isValidAggregate).collect(Collectors.toSet());
     } else {
-      facetsToAggregate = _defaultFacetFields;
+      facetsToAggregate = defaultFacetFields;
     }
     return facetsToAggregate.stream()
         .map(this::facetToAggregationBuilder)
@@ -79,13 +114,13 @@ public class AggregationQueryBuilder {
         !facets.isEmpty()
             && ((facets.size() == 1
                     && facets.get(0).startsWith(STRUCTURED_PROPERTY_MAPPING_FIELD + "."))
-                || _allFacetFields.containsAll(facets));
+                || allFacetFields.containsAll(facets));
     if (!isValid) {
       log.warn(
           String.format(
               "Requested facet for search filter aggregations that isn't part of the filters. "
                   + "Provided: %s; Available: %s",
-              inputFacet, _allFacetFields));
+              inputFacet, allFacetFields));
     }
     return isValid;
   }
@@ -122,11 +157,11 @@ public class AggregationQueryBuilder {
             facet.equalsIgnoreCase(INDEX_VIRTUAL_FIELD)
                 ? AggregationBuilders.terms(inputFacet)
                     .field(getAggregationField("_index"))
-                    .size(_configs.getMaxTermBucketSize())
+                    .size(configs.getMaxTermBucketSize())
                     .minDocCount(0)
                 : AggregationBuilders.terms(inputFacet)
                     .field(getAggregationField(facet))
-                    .size(_configs.getMaxTermBucketSize());
+                    .size(configs.getMaxTermBucketSize());
       }
       if (lastAggBuilder != null) {
         aggBuilder = aggBuilder.subAggregation(lastAggBuilder);
@@ -172,5 +207,366 @@ public class AggregationQueryBuilder {
           MISSING_SPECIAL_TYPE + AGGREGATION_SPECIAL_TYPE_DELIMITER + annotation.getFieldName());
     }
     return facetsFromAnnotation;
+  }
+
+  private String computeDisplayName(String name) {
+    if (getFacetToDisplayNames().containsKey(name)) {
+      return getFacetToDisplayNames().get(name);
+    } else if (name.contains(AGGREGATION_SEPARATOR_CHAR)) {
+      return Arrays.stream(name.split(AGGREGATION_SEPARATOR_CHAR))
+          .map(i -> getFacetToDisplayNames().get(i))
+          .collect(Collectors.joining(AGGREGATION_SEPARATOR_CHAR));
+    }
+    return name;
+  }
+
+  List<AggregationMetadata> extractAggregationMetadata(
+      @Nonnull SearchResponse searchResponse, @Nullable Filter filter) {
+    final List<AggregationMetadata> aggregationMetadataList = new ArrayList<>();
+    if (searchResponse.getAggregations() == null) {
+      return addFiltersToAggregationMetadata(aggregationMetadataList, filter);
+    }
+    for (Map.Entry<String, Aggregation> entry :
+        searchResponse.getAggregations().getAsMap().entrySet()) {
+      if (entry.getValue() instanceof ParsedTerms) {
+        processTermAggregations(entry, aggregationMetadataList);
+      }
+      if (entry.getValue() instanceof ParsedMissing) {
+        processMissingAggregations(entry, aggregationMetadataList);
+      }
+    }
+    return addFiltersToAggregationMetadata(aggregationMetadataList, filter);
+  }
+
+  private void processTermAggregations(
+      final Map.Entry<String, Aggregation> entry,
+      final List<AggregationMetadata> aggregationMetadataList) {
+    final Map<String, Long> oneTermAggResult =
+        extractTermAggregations(
+            (ParsedTerms) entry.getValue(), entry.getKey().equals(INDEX_VIRTUAL_FIELD));
+    if (oneTermAggResult.isEmpty()) {
+      return;
+    }
+    final AggregationMetadata aggregationMetadata =
+        new AggregationMetadata()
+            .setName(entry.getKey())
+            .setDisplayName(computeDisplayName(entry.getKey()))
+            .setAggregations(new LongMap(oneTermAggResult))
+            .setFilterValues(
+                new FilterValueArray(
+                    SearchUtil.convertToFilters(oneTermAggResult, Collections.emptySet())));
+    aggregationMetadataList.add(aggregationMetadata);
+  }
+
+  /**
+   * Adds nested sub-aggregation values to the aggregated results
+   *
+   * @param aggs The aggregations to traverse. Could be null (base case)
+   * @return A map from names to aggregation count values
+   */
+  @Nonnull
+  private static Map<String, Long> recursivelyAddNestedSubAggs(@Nullable Aggregations aggs) {
+    final Map<String, Long> aggResult = new HashMap<>();
+
+    if (aggs != null) {
+      for (Map.Entry<String, Aggregation> entry : aggs.getAsMap().entrySet()) {
+        if (entry.getValue() instanceof ParsedTerms) {
+          recurseTermsAgg((ParsedTerms) entry.getValue(), aggResult, false);
+        } else if (entry.getValue() instanceof ParsedMissing) {
+          recurseMissingAgg((ParsedMissing) entry.getValue(), aggResult);
+        } else {
+          throw new UnsupportedOperationException(
+              "Unsupported aggregation type: " + entry.getValue().getClass().getName());
+        }
+      }
+    }
+    return aggResult;
+  }
+
+  private static void recurseTermsAgg(
+      ParsedTerms terms, Map<String, Long> aggResult, boolean includeZeroes) {
+    List<? extends Terms.Bucket> bucketList = terms.getBuckets();
+    bucketList.forEach(bucket -> processTermBucket(bucket, aggResult, includeZeroes));
+  }
+
+  private static void processTermBucket(
+      Terms.Bucket bucket, Map<String, Long> aggResult, boolean includeZeroes) {
+    String key = bucket.getKeyAsString();
+    // Gets filtered sub aggregation doc count if exist
+    Map<String, Long> subAggs = recursivelyAddNestedSubAggs(bucket.getAggregations());
+    subAggs.forEach(
+        (entryKey, entryValue) ->
+            aggResult.put(
+                String.format("%s%s%s", key, AGGREGATION_SEPARATOR_CHAR, entryKey), entryValue));
+    long docCount = bucket.getDocCount();
+    if (includeZeroes || docCount > 0) {
+      aggResult.put(key, docCount);
+    }
+  }
+
+  private static void recurseMissingAgg(ParsedMissing missing, Map<String, Long> aggResult) {
+    Map<String, Long> subAggs = recursivelyAddNestedSubAggs(missing.getAggregations());
+    subAggs.forEach(
+        (key, value) ->
+            aggResult.put(
+                String.format("%s%s%s", missing.getName(), AGGREGATION_SEPARATOR_CHAR, key),
+                value));
+    long docCount = missing.getDocCount();
+    if (docCount > 0) {
+      aggResult.put(missing.getName(), docCount);
+    }
+  }
+
+  /**
+   * Extracts term aggregations give a parsed term.
+   *
+   * @param terms an abstract parse term, input can be either ParsedStringTerms ParsedLongTerms
+   * @return a map with aggregation key and corresponding doc counts
+   */
+  @Nonnull
+  private static Map<String, Long> extractTermAggregations(
+      @Nonnull ParsedTerms terms, boolean includeZeroes) {
+
+    final Map<String, Long> aggResult = new HashMap<>();
+    recurseTermsAgg(terms, aggResult, includeZeroes);
+
+    return aggResult;
+  }
+
+  /** Injects the missing conjunctive filters into the aggregations list. */
+  public List<AggregationMetadata> addFiltersToAggregationMetadata(
+      @Nonnull final List<AggregationMetadata> originalMetadata, @Nullable final Filter filter) {
+    if (filter == null) {
+      return originalMetadata;
+    }
+    if (filter.getOr() != null) {
+      addOrFiltersToAggregationMetadata(filter.getOr(), originalMetadata);
+    } else if (filter.getCriteria() != null) {
+      addCriteriaFiltersToAggregationMetadata(filter.getCriteria(), originalMetadata);
+    }
+    return originalMetadata;
+  }
+
+  void addOrFiltersToAggregationMetadata(
+      @Nonnull final ConjunctiveCriterionArray or,
+      @Nonnull final List<AggregationMetadata> originalMetadata) {
+    for (ConjunctiveCriterion conjunction : or) {
+      // For each item in the conjunction, inject an empty aggregation if necessary
+      addCriteriaFiltersToAggregationMetadata(conjunction.getAnd(), originalMetadata);
+    }
+  }
+
+  private void addCriteriaFiltersToAggregationMetadata(
+      @Nonnull final CriterionArray criteria,
+      @Nonnull final List<AggregationMetadata> originalMetadata) {
+    for (Criterion criterion : criteria) {
+      addCriterionFiltersToAggregationMetadata(criterion, originalMetadata);
+    }
+  }
+
+  private void addCriterionFiltersToAggregationMetadata(
+      @Nonnull final Criterion criterion,
+      @Nonnull final List<AggregationMetadata> aggregationMetadata) {
+
+    // We should never see duplicate aggregation for the same field in aggregation metadata list.
+    final Map<String, AggregationMetadata> aggregationMetadataMap =
+        aggregationMetadata.stream()
+            .collect(Collectors.toMap(AggregationMetadata::getName, agg -> agg));
+
+    // Map a filter criterion to a facet field (e.g. domains.keyword -> domains)
+    final String finalFacetField = toFacetField(criterion.getField());
+
+    if (finalFacetField == null) {
+      log.warn(
+          String.format(
+              "Found invalid filter field for entity search. Invalid or unrecognized facet %s",
+              criterion.getField()));
+      return;
+    }
+
+    // We don't want to add urn filters to the aggregations we return as a sidecar to search
+    // results.
+    // They are automatically added by searchAcrossLineage and we dont need them to show up in the
+    // filter panel.
+    if (finalFacetField.equals(URN_FILTER)) {
+      return;
+    }
+
+    if (aggregationMetadataMap.containsKey(finalFacetField)) {
+      /*
+       * If we already have aggregations for the facet field, simply inject any missing values counts into the set.
+       * If there are no results for a particular facet value, it will NOT be in the original aggregation set returned by
+       * Elasticsearch.
+       */
+      AggregationMetadata originalAggMetadata = aggregationMetadataMap.get(finalFacetField);
+      if (criterion.hasValues()) {
+        criterion
+            .getValues()
+            .forEach(
+                value ->
+                    addMissingAggregationValueToAggregationMetadata(value, originalAggMetadata));
+      } else {
+        addMissingAggregationValueToAggregationMetadata(criterion.getValue(), originalAggMetadata);
+      }
+    } else {
+      /*
+       * If we do not have ANY aggregation for the facet field, then inject a new aggregation metadata object for the
+       * facet field.
+       * If there are no results for a particular facet, it will NOT be in the original aggregation set returned by
+       * Elasticsearch.
+       */
+      aggregationMetadata.add(
+          buildAggregationMetadata(
+              finalFacetField,
+              getFacetToDisplayNames().getOrDefault(finalFacetField, finalFacetField),
+              new LongMap(
+                  criterion.getValues().stream().collect(Collectors.toMap(i -> i, i -> 0L))),
+              new FilterValueArray(
+                  criterion.getValues().stream()
+                      .map(value -> createFilterValue(value, 0L, true))
+                      .collect(Collectors.toList()))));
+    }
+  }
+
+  private void addMissingAggregationValueToAggregationMetadata(
+      @Nonnull final String value, @Nonnull final AggregationMetadata originalMetadata) {
+    if (originalMetadata.getAggregations().entrySet().stream()
+            .noneMatch(entry -> value.equals(entry.getKey()))
+        || originalMetadata.getFilterValues().stream()
+            .noneMatch(entry -> entry.getValue().equals(value))) {
+      // No aggregation found for filtered value -- inject one!
+      originalMetadata.getAggregations().put(value, 0L);
+      originalMetadata.getFilterValues().add(createFilterValue(value, 0L, true));
+    }
+  }
+
+  private AggregationMetadata buildAggregationMetadata(
+      @Nonnull final String facetField,
+      @Nonnull final String displayName,
+      @Nonnull final LongMap aggValues,
+      @Nonnull final FilterValueArray filterValues) {
+    return new AggregationMetadata()
+        .setName(facetField)
+        .setDisplayName(displayName)
+        .setAggregations(aggValues)
+        .setFilterValues(filterValues);
+  }
+
+  private List<Pair<String, Pair<String, String>>> getFacetFieldDisplayNameFromAnnotation(
+      @Nonnull EntitySpec entitySpec, @Nonnull final SearchableAnnotation annotation) {
+    final List<Pair<String, Pair<String, String>>> facetsFromAnnotation = new ArrayList<>();
+    // Case 1: Default Keyword field
+    if (annotation.isAddToFilters()) {
+      facetsFromAnnotation.add(
+          Pair.of(
+              annotation.getFieldName(),
+              Pair.of(entitySpec.getName(), annotation.getFilterName())));
+    }
+    // Case 2: HasX boolean field
+    if (annotation.isAddHasValuesToFilters() && annotation.getHasValuesFieldName().isPresent()) {
+      facetsFromAnnotation.add(
+          Pair.of(
+              annotation.getHasValuesFieldName().get(),
+              Pair.of(entitySpec.getName(), annotation.getHasValuesFilterName())));
+    }
+    return facetsFromAnnotation;
+  }
+
+  @WithSpan
+  public static Map<String, Long> extractAggregationsFromResponse(
+      @Nonnull SearchResponse searchResponse, @Nonnull String aggregationName) {
+    if (searchResponse.getAggregations() == null) {
+      return Collections.emptyMap();
+    }
+
+    Aggregation aggregation = searchResponse.getAggregations().get(aggregationName);
+    if (aggregation == null) {
+      return Collections.emptyMap();
+    }
+    if (aggregation instanceof ParsedTerms) {
+      return extractTermAggregations(
+          (ParsedTerms) aggregation, aggregationName.equals("_entityType"));
+    } else if (aggregation instanceof ParsedMissing) {
+      return Collections.singletonMap(
+          aggregation.getName(), ((ParsedMissing) aggregation).getDocCount());
+    }
+    throw new UnsupportedOperationException(
+        "Unsupported aggregation type: " + aggregation.getClass().getName());
+  }
+
+  /**
+   * Only used in aggregation queries, lazy load
+   *
+   * @return map of field name to facet display names
+   */
+  private Map<String, String> getFacetToDisplayNames() {
+    if (filtersToDisplayName == null) {
+      // Validate field names
+      Map<String, Set<Pair<String, Pair<String, String>>>> validateFieldMap =
+          entitySearchAnnotations.entrySet().stream()
+              .flatMap(
+                  entry ->
+                      entry.getValue().stream()
+                          .flatMap(
+                              annotation ->
+                                  getFacetFieldDisplayNameFromAnnotation(entry.getKey(), annotation)
+                                      .stream()))
+              .collect(Collectors.groupingBy(Pair::getFirst, Collectors.toSet()));
+      for (Map.Entry<String, Set<Pair<String, Pair<String, String>>>> entry :
+          validateFieldMap.entrySet()) {
+        if (entry.getValue().stream().map(i -> i.getSecond().getSecond()).distinct().count() > 1) {
+          Map<String, Set<Pair<String, String>>> displayNameEntityMap =
+              entry.getValue().stream()
+                  .map(Pair::getSecond)
+                  .collect(Collectors.groupingBy(Pair::getSecond, Collectors.toSet()));
+          throw new IllegalStateException(
+              String.format(
+                  "Facet field collision on field `%s`. Incompatible Display Name across entities. Multiple Display Names detected: %s",
+                  entry.getKey(), displayNameEntityMap));
+        }
+      }
+
+      filtersToDisplayName =
+          entitySearchAnnotations.entrySet().stream()
+              .flatMap(
+                  entry ->
+                      entry.getValue().stream()
+                          .flatMap(
+                              annotation ->
+                                  getFacetFieldDisplayNameFromAnnotation(entry.getKey(), annotation)
+                                      .stream()))
+              .collect(
+                  Collectors.toMap(Pair::getFirst, p -> p.getSecond().getSecond(), mapMerger()));
+      filtersToDisplayName.put(INDEX_VIRTUAL_FIELD, "Type");
+    }
+
+    return filtersToDisplayName;
+  }
+
+  private void processMissingAggregations(
+      final Map.Entry<String, Aggregation> entry,
+      final List<AggregationMetadata> aggregationMetadataList) {
+    ParsedMissing parsedMissing = (ParsedMissing) entry.getValue();
+    Long docCount = parsedMissing.getDocCount();
+    LongMap longMap = new LongMap();
+    longMap.put(entry.getKey(), docCount);
+    final AggregationMetadata aggregationMetadata =
+        new AggregationMetadata()
+            .setName(entry.getKey())
+            .setDisplayName(computeDisplayName(entry.getKey()))
+            .setAggregations(longMap)
+            .setFilterValues(
+                new FilterValueArray(SearchUtil.convertToFilters(longMap, Collections.emptySet())));
+    aggregationMetadataList.add(aggregationMetadata);
+  }
+
+  // If values are not equal, throw error
+  private BinaryOperator<String> mapMerger() {
+    return (s1, s2) -> {
+      if (!StringUtils.equals(s1, s2)) {
+        throw new IllegalStateException(String.format("Unable to merge values %s and %s", s1, s2));
+      }
+      return s1;
+    };
   }
 }
