+++ metadata-service/factories/src/main/java/com/linkedin/gms/factory/recommendation/candidatesource/DomainsCandidateSourceFactory.java
@@ -1,6 +1,7 @@
 package com.linkedin.gms.factory.recommendation.candidatesource;
 
 import com.linkedin.gms.factory.search.EntitySearchServiceFactory;
+import com.linkedin.metadata.models.registry.EntityRegistry;
 import com.linkedin.metadata.recommendation.candidatesource.DomainsCandidateSource;
 import com.linkedin.metadata.search.EntitySearchService;
 import javax.annotation.Nonnull;
@@ -20,7 +21,7 @@ public class DomainsCandidateSourceFactory {
 
   @Bean(name = "domainsCandidateSource")
   @Nonnull
-  protected DomainsCandidateSource getInstance() {
-    return new DomainsCandidateSource(entitySearchService);
+  protected DomainsCandidateSource getInstance(final EntityRegistry entityRegistry) {
+    return new DomainsCandidateSource(entitySearchService, entityRegistry);
   }
 }
